refactor blaster module file name
