+++ synphage/assets/blaster/n_blaster.py
