+++ test/unit/blaster_asset/test_fasta_conversion.py
@@ -1,7 +1,7 @@
 from pathlib import PosixPath
 from dagster import materialize_to_memory, build_asset_context, asset
 
-from synphage.assets.blaster.blaster import genbank_to_fasta
+from synphage.assets.blaster.n_blaster import genbank_to_fasta
 
 
 TEST_DATA_GB_DIR = "test/fixtures/assets_testing_folder/blasting/genbank/"
